+++ deepinv/datasets/__init__.py
@@ -5,3 +5,7 @@ from .urban100 import Urban100HR
 from .set14 import Set14HR
 from .cbsd68 import CBSD68
 from .fastmri import FastMRISliceDataset
+from .lidc_idri import LidcIdriSliceDataset
+from .flickr2k import Flickr2kHR
+from .lsdir import LsdirHR
+from .fmd import FMD