+++ deepinv/loss/measplit.py
@@ -1,14 +1,16 @@
+from typing import Optional
+from warnings import warn
 import torch
-from deepinv.physics import Inpainting
+from deepinv.physics import Inpainting, Denoising, Physics
 from deepinv.loss.loss import Loss
-from deepinv.physics.generator import BernoulliMaskGenerator
+from deepinv.physics.generator import PhysicsGenerator, BernoulliSplittingMaskGenerator
 
 
 class SplittingLoss(Loss):
     r"""
     Measurement splitting loss.
 
-    Splits the measurement and forward operator (of size :math:`m`)
+    Splits the measurement and forward operator :math:`\forw{}` (of size :math:`m`)
     into two smaller pairs  :math:`(y_1,A_1)` (of size :math:`m_1`) and  :math:`(y_2,A_2)` (of size :math:`m_2`) ,
     to compute the self-supervised loss:
 
@@ -16,7 +18,14 @@ class SplittingLoss(Loss):
 
         \frac{m}{m_2}\| y_2 - A_2 \inversef{y_1}{A_1}\|^2
 
-    where :math:`R` is the trainable network. See https://pubmed.ncbi.nlm.nih.gov/32614100/.
+    where :math:`R` is the trainable network, :math:`A_1 = M_1 \forw{}, A_2 = M_2 \forw{}`, and :math:`M_i` are randomly generated masks (i.e. diagonal matrices) such that :math:`M_1+M_2=\mathbb{I}_m`.
+
+    .. note::
+
+        If the forward operator has its own subsampling mask :math:`M_{\forw{}}`, e.g. :class:`deepinv.physics.Inpainting` or :class:`deepinv.physics.MRI`,
+        the splitting masks will be subsets of the physics' mask such that :math:`M_1+M_2=M_{\forw{}}`
+
+    This loss was used in SSDU for MRI in `Yaman et al. Self-supervised learning of physics-guided reconstruction neural networks without fully sampled reference data <https://pubmed.ncbi.nlm.nih.gov/32614100/>`_
 
     By default, the error is computed using the MSE metric, however any other metric (e.g., :math:`\ell_1`)
     can be used as well.
@@ -29,14 +38,24 @@ class SplittingLoss(Loss):
 
         To obtain the best test performance, the trained model should be averaged at test time
         over multiple realizations of the splitting, i.e.
-        :math:`\hat{x} = \frac{1}{N}\sum_{i=1}^N \inversef{y_1^{(i)}}{A_1^{(i)}}`.
+        :math:`\hat{x} = \frac{1}{N}\sum_{i=1}^N \inversef{y_1^{(i)}}{A_1^{(i)}}`. To disable this, set ``MC_samples=1``.
+
+    .. note::
+
+        To disable measurement splitting (and use the full input) at evaluation time, set ``eval_split_input=True``. This is done in `SSDU <https://pubmed.ncbi.nlm.nih.gov/32614100/>`_.
 
     :param torch.nn.Module metric: metric used for computing data consistency,
         which is set as the mean squared error by default.
     :param float split_ratio: splitting ratio, should be between 0 and 1. The size of :math:`y_1` increases
         with the splitting ratio.
     :param deepinv.physics.generator.PhysicsGenerator, None mask_generator: function to generate the mask. If
-        None, the :class:`deepinv.physics.generator.BernoulliMaskGenerator` is used.
+        None, the :class:`deepinv.physics.generator.BernoulliSplittingMaskGenerator` is used.
+    :param int MC_samples: Number of samples used for averaging. Must be greater than 0.
+    :param bool eval_split_input: if True, perform input measurement splitting during evaluation. If False, use full measurement at eval (no MC samples are performed and eval_split_output will have no effect)
+    :param bool eval_split_output: at evaluation time, pass the output through the output mask too.
+        i.e. :math:`(\sum_{j=1}^N M_2^{(j)})^{-1} \sum_{i=1}^N M_2^{(i)} \inversef{y_1^{(i)}}{A_1^{(i)}}`.
+        Only valid when y is same domain (and dimension) as x. Defaults to False.
+    :param bool pixelwise: if True, create pixelwise splitting masks i.e. zero all channels simultaneously.
 
     |sep|
 
@@ -46,8 +65,8 @@ class SplittingLoss(Loss):
     >>> import deepinv as dinv
     >>> physics = dinv.physics.Inpainting(tensor_size=(1, 8, 8), mask=0.5)
     >>> model = dinv.models.MedianFilter()
-    >>> loss = dinv.loss.SplittingLoss(split_ratio=0.9)
-    >>> model = loss.adapt_model(model, MC_samples=2) # important step!
+    >>> loss = dinv.loss.SplittingLoss(split_ratio=0.9, MC_samples=2)
+    >>> model = loss.adapt_model(model) # important step!
     >>> x = torch.ones((1, 1, 8, 8))
     >>> y = physics(x)
     >>> x_net = model(y, physics, update_parameters=True) # save random mask in forward pass
@@ -58,12 +77,25 @@ class SplittingLoss(Loss):
 
     """
 
-    def __init__(self, metric=torch.nn.MSELoss(), split_ratio=0.9, mask_generator=None):
+    def __init__(
+        self,
+        metric=torch.nn.MSELoss(),
+        split_ratio: float = 0.9,
+        mask_generator: Optional[PhysicsGenerator] = None,
+        MC_samples=5,
+        eval_split_input=True,
+        eval_split_output=False,
+        pixelwise=True,
+    ):
         super().__init__()
         self.name = "ms"
         self.metric = metric
         self.mask_generator = mask_generator
         self.split_ratio = split_ratio
+        self.MC_samples = MC_samples
+        self.eval_split_input = eval_split_input
+        self.eval_split_output = eval_split_output
+        self.pixelwise = pixelwise
 
     def forward(self, x_net, y, physics, model, **kwargs):
         r"""
@@ -74,17 +106,15 @@ class SplittingLoss(Loss):
         :param torch.nn.Module model: Reconstruction function.
         :return: (torch.Tensor) loss.
         """
-        tsize = y.size()[1:]
         mask = model.get_mask()
 
         # create inpainting masks
-        mask2 = 1.0 - mask
-        inp2 = Inpainting(
-            tsize, mask=mask2, device=y.device, noise_model=physics.noise_model
-        )
+        mask2 = getattr(physics, "mask", 1.0) - mask
+        inp2 = Inpainting(y.size()[1:], mask=mask2, device=y.device)
 
         # concatenate operators
         physics2 = inp2 * physics  # A_2 = (I-P)*A
+        physics2.noise_model = physics.noise_model
 
         # divide measurements
         y2 = inp2.A(y)
@@ -95,7 +125,7 @@ class SplittingLoss(Loss):
 
         return loss_ms
 
-    def adapt_model(self, model, MC_samples=5):
+    def adapt_model(self, model: torch.nn.Module, MC_samples=None):
         r"""
         Apply random splitting to input.
 
@@ -111,52 +141,129 @@ class SplittingLoss(Loss):
         randomly splitting the measurements :math:`y` and operator :math:`A`.
         During training (i.e. when ``model.train()``), we use only one sample, i.e. :math:`N=1`
         for computational efficiency, whereas at test time, we use multiple samples for better performance.
+        For other parameters that control how splitting is applied, see the class parameters.
 
         :param torch.nn.Module model: Reconstruction model.
-        :param int MC_samples: Number of samples used for averaging.
+        :param int MC_samples: deprecated. Pass ``MC_samples`` at class initialisation instead.
         :return: (torch.nn.Module) Model modified for evaluation.
         """
+        if MC_samples is not None:
+            warn(
+                "MC_samples parameter is deprecated. Pass MC_samples at init: SplittingLoss(MC_samples=...)"
+            )
 
         if isinstance(model, SplittingModel):
             return model
         else:
             return SplittingModel(
-                model, self.split_ratio, self.mask_generator, MC_samples
+                model,
+                split_ratio=self.split_ratio,
+                mask_generator=self.mask_generator,
+                MC_samples=self.MC_samples,
+                eval_split_input=self.eval_split_input,
+                eval_split_output=self.eval_split_output,
+                pixelwise=self.pixelwise,
             )
 
 
 class SplittingModel(torch.nn.Module):
-    def __init__(self, model, split_ratio, mask_generator, MC_samples):
+    """
+    Model wrapper when using SplittingLoss.
+
+    Performs input splitting during forward pass. At evaluation, perform forward passes for multiple realisations of splitting mask and average.
+
+    For parameters, see :class:`deepinv.loss.SplittingLoss`
+    """
+
+    def __init__(
+        self,
+        model,
+        split_ratio,
+        mask_generator,
+        MC_samples,
+        eval_split_input,
+        eval_split_output,
+        pixelwise,
+    ):
         super().__init__()
         self.model = model
         self.split_ratio = split_ratio
         self.MC_samples = MC_samples
         self.mask = 0
         self.mask_generator = mask_generator
+        self.eval_split_input = eval_split_input
+        self.eval_split_output = eval_split_output
+        self.pixelwise = pixelwise
 
-    def forward(self, y, physics, update_parameters=False):
-        MC = 1 if self.training else self.MC_samples
-        tsize = y.size()[1:]
+    def forward(
+        self, y: torch.Tensor, physics: Physics, update_parameters: bool = False
+    ):
+        """Adapted model forward pass for input splitting. During training, only one splitting realisation is performed for computational efficiency."""
         out = 0
 
         if self.mask_generator is None:
-            self.mask_generator = BernoulliMaskGenerator(
-                tensor_size=tsize, split_ratio=self.split_ratio, device=y.device
+            self.mask_generator = BernoulliSplittingMaskGenerator(
+                tensor_size=y.size()[1:],
+                split_ratio=self.split_ratio,
+                pixelwise=self.pixelwise,
+                device=y.device,
             )
 
-        inp = Inpainting(
-            tsize, mask=0.0, device=y.device, noise_model=physics.noise_model
-        )
+        inp = Inpainting(y.size()[1:], device=y.device)
+
         with torch.set_grad_enabled(self.training):
-            for i in range(MC):
-                mask = self.mask_generator.step(y.size(0))
-                y1 = inp.A(y, **mask)
-                physics1 = inp * physics  # A_1 = P*A
-                out += self.model(y1, physics1)
+
+            if not self.eval_split_input and not self.training:
+                return self.model(y, physics)
+            elif self.eval_split_output and self.eval_split_input and not self.training:
+                return self._forward_split_output(y, physics)
+
+            else:
+                MC_samples = 1 if self.training else self.MC_samples
+
+                for _ in range(MC_samples):
+                    # Perform input masking
+                    mask = self.mask_generator.step(
+                        y.size(0), input_mask=getattr(physics, "mask", None)
+                    )
+                    y1 = inp.A(y, **mask)
+                    physics1 = inp * physics  # A_1 = P*A
+                    physics1.noise_model = physics.noise_model
+
+                    # Forward pass
+                    out += self.model(y1, physics1) / MC_samples
 
             if self.training and update_parameters:
                 self.mask = mask["mask"]
-            out = out / MC
+
+        return out
+
+    def _forward_split_output(self, y: torch.Tensor, physics: Physics):
+        """Perform splitting at model output too, only at eval time"""
+        out = 0
+        normaliser = torch.zeros_like(y)
+        inp = Inpainting(y.size()[1:], device=y.device)
+
+        for _ in range(self.MC_samples):
+            # Perform input masking
+            mask = self.mask_generator.step(
+                y.size(0), input_mask=getattr(physics, "mask", None)
+            )
+            y1 = inp.A(y, **mask)
+            physics1 = inp * physics  # A_1 = P*A
+            physics1.noise_model = physics.noise_model
+
+            # Forward pass
+            x_hat = self.model(y1, physics1)
+
+            # Output masking
+            mask2 = getattr(physics, "mask", 1.0) - mask["mask"]
+            inp2 = Inpainting(y.size()[1:], mask=mask2, device=y.device)
+            out += inp2.A(x_hat)
+            normaliser += mask2
+
+        out[normaliser != 0] /= normaliser[normaliser != 0]
+
         return out
 
     def get_mask(self):